# src/swagger.json
{
	"components": {
...
		"schemas": {
...
			"ErrorResponse": {
				"properties": {
					"name": {
						"type": "string"
					},
					"message": {
						"type": "string"
					},
					"errorUUID": {
						"type": "string"
					}
				},
				"required": [
					"name",
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"EpisodeResponse": {
				"properties": {
					"episode_id": {
						"type": "number",
						"format": "double"
					},
					"season": {
						"type": "number",
						"format": "double"
					},
					"episode": {
						"type": "number",
						"format": "double"
					},
					"netflix_id": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"episode_id",
					"season",
					"episode",
					"netflix_id"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"MovieNameResponse": {
				"properties": {
					"movie_name_id": {
						"type": "number",
						"format": "double"
					},
					"lang": {
						"type": "string"
					},
					"content": {
						"type": "string"
					}
				},
				"required": [
					"movie_name_id",
					"lang",
					"content"
				],
				"type": "object",
				"additionalProperties": false
			},
			"MovieResponse": {
				"properties": {
					"movie_id": {
						"type": "number",
						"format": "double"
					},
					"lang": {
						"type": "string"
					},
					"type": {
						"type": "string",
						"enum": [
							"movie",
							"serie"
						]
					},
					"netflix_id": {
						"type": "number",
						"format": "double"
					},
					"minutes": {
						"type": "number",
						"format": "double"
					},
					"movie_names": {
						"items": {
							"$ref": "#/components/schemas/MovieNameResponse"
						},
						"type": "array"
					},
					"episodes": {
						"items": {
							"$ref": "#/components/schemas/EpisodeResponse"
						},
						"type": "array"
					}
				},
				"required": [
					"movie_id",
					"lang",
					"type",
					"minutes",
					"movie_names",
					"episodes"
				],
				"type": "object",
				"additionalProperties": false
			},
			"LyricSentenceResponse": {
				"properties": {
					"sentence_id": {
						"type": "number",
						"format": "double"
					},
					"lang": {
						"type": "string"
					},
					"content": {
						"type": "string"
					}
				},
				"required": [
					"sentence_id",
					"lang",
					"content"
				],
				"type": "object",
				"additionalProperties": false
			},
			"LyricResponse": {
				"properties": {
					"lyric_id": {
						"type": "number",
						"format": "double"
					},
					"movie": {
						"$ref": "#/components/schemas/MovieResponse"
					},
					"episode": {
						"allOf": [
							{
								"$ref": "#/components/schemas/EpisodeResponse"
							}
						],
						"nullable": true
					},
					"seconds": {
						"type": "number",
						"format": "double"
					},
					"quality": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"sentences": {
						"items": {
							"$ref": "#/components/schemas/LyricSentenceResponse"
						},
						"type": "array"
					}
				},
				"required": [
					"lyric_id",
					"movie",
					"episode",
					"seconds",
					"quality",
					"sentences"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PostLyricRequest": {
				"properties": {
					"seconds": {
						"type": "number",
						"format": "double"
					},
					"movieId": {
						"type": "number",
						"format": "double"
					},
					"quality": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"episodeId": {
						"type": "number",
						"format": "double",
						"nullable": true
					}
				},
				"required": [
					"seconds",
					"movieId",
					"quality",
					"episodeId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PutLyricRequest": {
				"properties": {
					"seconds": {
						"type": "number",
						"format": "double"
					},
					"movieId": {
						"type": "number",
						"format": "double"
					},
					"quality": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"episodeId": {
						"type": "number",
						"format": "double",
						"nullable": true
					}
				},
				"required": [
					"seconds",
					"movieId",
					"quality",
					"episodeId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PostLyricSentenceRequest": {
				"properties": {
					"lyric_id": {
						"type": "number",
						"format": "double"
					},
					"lang": {
						"type": "string"
					},
					"content": {
						"type": "string"
					}
				},
				"required": [
					"lyric_id",
					"lang",
					"content"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PutLyricSentenceRequest": {
				"properties": {
					"lang": {
						"type": "string"
					},
					"content": {
						"type": "string"
					}
				},
				"required": [
					"lang",
					"content"
				],
				"type": "object",
				"additionalProperties": false
			},
			"MovieType": {
				"enum": [
					"only_movies",
					"only_series"
				],
				"type": "string"
			},
			"PostMovieRequest": {
				"properties": {
					"lang": {
						"type": "string"
					},
					"netflix_id": {
						"type": "number",
						"format": "double"
					},
					"minutes": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"lang",
					"minutes"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PutMovieRequest": {
				"properties": {
					"lang": {
						"type": "string"
					},
					"netflix_id": {
						"type": "number",
						"format": "double"
					},
					"minutes": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"lang",
					"netflix_id",
					"minutes"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PostMovieNameRequest": {
				"properties": {
					"lang": {
						"type": "string"
					},
					"content": {
						"type": "string"
					},
					"movieId": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"lang",
					"content",
					"movieId"
				],
				"type": "object",
				"additionalProperties": false
			},
			"PutMovieNameRequest": {
				"properties": {
					"lang": {
						"type": "string"
					},
					"content": {
						"type": "string"
					}
				},
				"required": [
					"lang",
					"content"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"api_key": {
				"type": "apiKey",
				"name": "Authorization",
				"in": "header",
				"scopes": {
					"contributor": "It allows you to modify (Create, Update, Delete) the content of the application. E.g. Movies, Languages etc.",
					"client": "It allows you to access application data. This permission should be the only one for the end user.",
					"admin": "It allows you to add new keys and permissions. It also allows you to assign permissions keys."
				}
			}
		}
	},
	"info": {
		"title": "lyricated-api",
		"version": "v2 javascript",
		"description": "API for Lyricated app.",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/lyric/new": {
			"post": {
				"operationId": "CreateLyric",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LyricResponse"
								}
							}
						}
					},
					"422": {
						"description": "Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Lyric"
				],
				"security": [
					{
						"api_key": [
							"contributor"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/PostLyricRequest"
							}
						}
					}
				}
			}
		},
...
		"/lyric/{id}": {
			"get": {
				"operationId": "GetLyricById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LyricResponse"
								}
							}
						}
					},
					"404": {
						"description": "Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Lyric"
				],
				"security": [
					{
						"api_key": [
							"client"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "PutLyric",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LyricResponse"
								}
							}
						}
					},
					"404": {
						"description": "Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Lyric"
				],
				"security": [
					{
						"api_key": [
							"contributor"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/PutLyricRequest"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteLyric",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LyricResponse"
								}
							}
						}
					},
					"422": {
						"description": "Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Lyric"
				],
				"security": [
					{
						"api_key": [
							"contributor"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/lyric-sentence/new": {
			"post": {
				"operationId": "CreateLyricSentence",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LyricSentenceResponse"
								}
							}
						}
					},
					"400": {
						"description": "Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"tags": [
					"LyricSentence"
				],
				"security": [
					{
						"api_key": [
							"contributor"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/PostLyricSentenceRequest"
							}
						}
					}
				}
			}
		},
		"/lyric-sentence/{id}": {
			"get": {
				"operationId": "GetLyricSentence",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LyricSentenceResponse"
								}
							}
						}
					},
					"400": {
						"description": "Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"tags": [
					"LyricSentence"
				],
				"security": [
					{
						"api_key": [
							"client"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "PutLyricSentence",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LyricSentenceResponse"
								}
							}
						}
					},
					"400": {
						"description": "Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"tags": [
					"LyricSentence"
				],
				"security": [
					{
						"api_key": [
							"contributor"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/PutLyricSentenceRequest"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteMovie",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/LyricSentenceResponse"
								}
							}
						}
					},
					"400": {
						"description": "Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"tags": [
					"LyricSentence"
				],
				"security": [
					{
						"api_key": [
							"contributor"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/movie/find": {
			"get": {
				"operationId": "GetMovies",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"properties": {
										"movies": {
											"items": {
												"$ref": "#/components/schemas/MovieResponse"
											},
											"type": "array"
										}
									},
									"required": [
										"movies"
									],
									"type": "object"
								}
							}
						}
					}
				},
				"tags": [
					"Movie"
				],
				"security": [
					{
						"api_key": [
							"client"
						]
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "type",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/MovieType"
						}
					}
				]
			}
		},
		"/movie/new": {
			"post": {
				"operationId": "CreateMovie",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MovieResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Movie"
				],
				"security": [
					{
						"api_key": [
							"contributor"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/PostMovieRequest"
							}
						}
					}
				}
			}
		},
		"/movie/{id}": {
			"get": {
				"operationId": "GetMovie",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MovieResponse"
								}
							}
						}
					},
					"404": {
						"description": "Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Movie"
				],
				"security": [
					{
						"api_key": [
							"client"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "PutMovie",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MovieResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Movie"
				],
				"security": [
					{
						"api_key": [
							"contributor"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/PutMovieRequest"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteMovie",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MovieResponse"
								}
							}
						}
					},
					"400": {
						"description": "Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Movie"
				],
				"security": [
					{
						"api_key": [
							"contributor"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/movie-name/new": {
			"post": {
				"operationId": "PostMovieName",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MovieNameResponse"
								}
							}
						}
					},
					"400": {
						"description": "Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"tags": [
					"MovieName"
				],
				"security": [
					{
						"api_key": [
							"contributor"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/PostMovieNameRequest"
							}
						}
					}
				}
			}
		},
		"/movie-name/{id}": {
			"get": {
				"operationId": "GetMovieName",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MovieNameResponse"
								}
							}
						}
					},
					"400": {
						"description": "Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"tags": [
					"MovieName"
				],
				"security": [
					{
						"api_key": [
							"client"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			},
			"put": {
				"operationId": "PutMovieName",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MovieNameResponse"
								}
							}
						}
					},
					"400": {
						"description": "Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"tags": [
					"MovieName"
				],
				"security": [
					{
						"api_key": [
							"contributor"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/PutMovieNameRequest"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteMovieName",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/MovieNameResponse"
								}
							}
						}
					},
					"400": {
						"description": "Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				},
				"tags": [
					"MovieName"
				],
				"security": [
					{
						"api_key": [
							"contributor"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}